+++ tests/test_lexbor.py
@@ -1,0 +1,20 @@
+"""Tests for functionality that is only supported by lexbor backend."""
+
+from selectolax.lexbor import LexborHTMLParser
+
+
+def test_reads_inner_html():
+    html = """<div id="main"><div>Hi</div><div id="updated">2025-09-27</div></div>"""
+    parser = LexborHTMLParser(html)
+    actual = parser.css_first("#main").inner_html
+    expected = """<div>Hi</div><div id="updated">2025-09-27</div>"""
+    assert actual == expected
+
+
+def test_sets_inner_html():
+    html = """<div id="main"><div>Hi</div><div id="updated">2025-09-27</div></div>"""
+    parser = LexborHTMLParser(html)
+    expected = "<span>Test</span>"
+    parser.css_first("#main").inner_html = "<span>Test</span>"
+    actual = parser.css_first("#main").inner_html
+    assert actual == expected