+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Fix docstring for `css_first` method
 - Fix memory leak in `merge_text_nodes` for lexbor backend
 - Update lexbor backend
+- Add `.inner_html` property. Allows to get and set inner HTML of a node.
 
 
 ## Version 0.3.34